made sure ceil got int

+++ plot_snana.py
@@ -88,7 +88,7 @@ def plot_spec(cid,bin_size,base_name):
 	if len(np.unique(sn['tobs']))>1:
 		figs=[]
 		m=0
-		for nfig in range(math.ceil(len(np.unique(sn['tobs']))/4.)):
+		for nfig in range(int(math.ceil(len(np.unique(sn['tobs']))/4.))):
 			fig,ax=plt.subplots(nrows=min(len(np.unique(sn['tobs'])),4),ncols=1,figsize=(8,8),sharex=True)
 			ax[0].set_title('SN%s'%cid[0],fontsize=16)
 			for j in range(min(len(np.unique(sn['tobs']))[m:],4)):
@@ -169,7 +169,7 @@ def plot_lc(cid,base_name):
                         [x for x in np.unique(sn['filter']) if x not in __band_order__])
 	
 	j=0
-	for nfig in range(math.ceil(rows/4.)): 
+	for nfig in range(int(math.ceil(rows/4.))): 
 		fig,ax=plt.subplots(nrows=min(rows,4),ncols=1,figsize=(8,8),sharex=True)
 		ax[0].set_title('SN%s'%cid[0],fontsize=16)
